refactor(idx): create `signInControllers` to handle sign in behaviour for idx api
